model that inbody:... intitle:... etc can have either simple term or phrase query as subterm
