+++ Cargo.toml
@@ -67,6 +67,7 @@ memmap = "0.7.0"
 memmap2 = "0.9.0"
 mime = "0.3.17"
 min-max-heap = "1.3.0"
+nom = "7.1.3"
 num_cpus = "1.15.0"
 once_cell = "1.13.1"
 proptest = "1.2.0"
